ENH: add real and imag with setters to DiscreteLpVector
Adding these requires mostly a utility to create the real
counterpart of a complex space with corresponding real data type.
This is implemented in the real_space and complex_space utils
in odl.util.utility.

Further changes:
- Get data types from numpy in base_ntuples.py instead of manually
  entering them into the type maps
- Accept an FnWeightingBase as weight argument in the initializers
  of Fn and CudaFn. That makes it easy to create a new space with
  the same weighting.
- move fast_1d_tensor_mult to numerics.py in odl.util
- remove some obsolete TODO comments in test files
