+++ odl/discr/lp_discr.py
@@ -461,6 +461,29 @@ class DiscreteLpVector(DiscretizationVector):
         """Axis ordering for array flattening."""
         return self.space.order
 
+    def real(self):
+        """Real part of this element."""
+        return real_space(self.space).element(self.asarray().real)
+
+    @real.setter
+    def real(self, newreal):
+        """Set the real part of this element to ``newreal``."""
+        newreal_flat = np.asarray(newreal, order=self.space.order).reshape(
+            -1, order=self.space.order)
+        self.ntuple.real = newreal_flat
+
+    @property
+    def imag(self):
+        """Imaginary part of this element."""
+        return real_space(self.space).element(self.asarray().imag)
+
+    @imag.setter
+    def imag(self, newimag):
+        """Set the imaginary part of this element to ``newimag``."""
+        newimag_flat = np.asarray(newimag, order=self.space.order).reshape(
+            -1, order=self.space.order)
+        self.ntuple.imag = newimag_flat
+
     def __setitem__(self, indices, values):
         """Set values of this vector.
