+++ odl/util/numerics.py
@@ -27,7 +27,7 @@ standard_library.install_aliases()
 import numpy as np
 
 
-__all__ = ('apply_on_boundary',)
+__all__ = ('apply_on_boundary', 'fast_1d_tensor_mult')
 
 
 def apply_on_boundary(array, func, only_once=True, which_boundaries=None,
@@ -178,6 +178,111 @@ def apply_on_boundary(array, func, only_once=True, which_boundaries=None,
         slices[ax] = slice(start, end)
 
     return out
+
+
+def fast_1d_tensor_mult(ndarr, onedim_arrs, axes=None):
+    """Fast multiplication of an n-dim array with an outer product.
+
+    This method implements the multiplication of an n-dimensional array
+    with an outer product of one-dimensional arrays, e.g.::
+
+        a = np.ones((10, 10, 10))
+        x = np.random.rand(10)
+        a *= x[:, None, None] * x[None, :, None] * x[None, None, :]
+
+    Basically, there are two ways to do such an operation:
+
+    1. First calculate the factor on the right-hand side and do one
+       "big" multiplication; or
+    2. Multiply by one factor at a time.
+
+    The procedure of building up the large factor in the first method
+    is relatively cheap if the number of 1d arrays is smaller than the
+    number of dimensions. For exactly n vectors, the second method is
+    faster, although it loops of the array ``a`` n times.
+
+    This implementation combines the two ideas into a hybrid scheme:
+
+    - If there are less 1d arrays than dimensions, choose 1.
+    - Otherwise, calculate the factor array for n-1 arrays
+      and multiply it to the large array. Finally, multiply with the
+      last 1d array.
+
+    The advantage of this approach is that it is memory-friendly and
+    loops over the big array only twice.
+
+    Parameters
+    ----------
+    ndarr : `numpy.ndarray`
+        Array to be multiplied with. Manipulations are done in-place.
+    onedim_arrs : sequence of array-like
+        One-dimensional arrays to be multiplied with ``ndarr``. The
+        sequence may not be longer than ``ndarr.ndim``.
+    axes : sequence of `int`, optional
+        Take the 1d transform along these axes. `None` corresponds to
+        the last ``len(onedim_arrs)`` axes, in ascending order.
+    """
+    if not isinstance(ndarr, np.ndarray):
+        raise TypeError('Expected a numpy.ndarray, got {!r}.'.format(ndarr))
+
+    if not onedim_arrs:
+        raise ValueError('No 1d arrays given.')
+
+    if axes is None:
+        axes = list(range(ndarr.ndim - len(onedim_arrs), ndarr.ndim))
+    elif len(axes) != len(onedim_arrs):
+        raise ValueError('There are {} 1d arrays, but {} axes entries.'
+                         ''.format(len(onedim_arrs), len(axes)))
+    else:
+        # Make axes positive
+        axes_ = np.array(axes, dtype=int)
+        axes_[axes_ < 0] += ndarr.ndim
+        axes = list(axes_)
+
+    if np.any(np.array(axes) >= ndarr.ndim) or np.any(np.array(axes) < 0):
+        raise ValueError('axes sequence contains out-of-bounds indices.')
+
+    # Make scalars 1d arrays and squeezable arrays 1d
+    alist = [np.atleast_1d(np.asarray(a).squeeze()) for a in onedim_arrs]
+    if any(a.ndim != 1 for a in alist):
+        raise ValueError('Only 1d arrays allowed.')
+
+    if len(axes) < ndarr.ndim:
+        # Make big factor array (start with 0d)
+        factor = np.array(1.0)
+        for ax, arr in zip(axes, alist):
+            # Meshgrid-style slice
+            slc = [None] * ndarr.ndim
+            slc[ax] = slice(None)
+            factor = factor * arr[slc]
+
+        ndarr *= factor
+
+    else:
+        # Hybrid approach
+
+        # Get the axis to spare for the final multiplication, the one
+        # with the largest stride.
+        axis_order = np.argsort(ndarr.strides)
+        last_ax = axis_order[-1]
+        last_arr = alist[axes.index(last_ax)]
+
+        # Build the semi-big array and multiply
+        factor = np.array(1.0)
+        for ax, arr in zip(axes, alist):
+            if ax == last_ax:
+                continue
+
+            slc = [None] * ndarr.ndim
+            slc[ax] = np.s_[:]
+            factor = factor * arr[slc]
+
+        ndarr *= factor
+
+        # Finally multiply by the remaining 1d array
+        slc = [None] * ndarr.ndim
+        slc[last_ax] = np.s_[:]
+        ndarr *= last_arr[slc]
 
 
 if __name__ == '__main__':
