+++ odl/util/utility.py
@@ -242,145 +242,6 @@ def is_complex_floating_dtype(dtype):
     return np.issubsctype(dtype, np.complexfloating)
 
 
-def default_dtype(impl, field):
-    """Return the default data type of different implementations."""
-    from odl.set.sets import RealNumbers
-    if impl == 'numpy':
-        if field == RealNumbers():
-            dtype = np.dtype('float64')
-        else:
-            dtype = np.dtype('complex128')
-    elif impl == 'cuda':
-        if field == RealNumbers():
-            dtype = np.dtype('float32')
-        else:
-            raise NotImplementedError('complex data types not supported in '
-                                      'CUDA.')
-            # dtype = np.dtype('complex64')
-    else:
-        raise ValueError("'impl '{}' not understood.".format(impl))
-
-    return dtype
-
-
-def complex_space(space):
-    """Return the complex counterpart of a given space.
-
-    Parameters
-    ----------
-    space : `LinearSpace`
-        Template space to be converted into its complex counterpart.
-        If ``space`` is already a complex space, a copy of it is
-        returned. Supported types of spaces: `FunctionSpace`, `FnBase`,
-        `DiscreteLp`.
-
-    Returns
-    -------
-    cspace : `LinearSpace`
-        Space of the same type and all parameters equal except field
-        and data type (if applicable), which are mapped to their complex
-        counterparts.
-
-    Examples
-    --------
-    >>> from odl import uniform_discr
-    >>> rspace = uniform_discr(0, 1, 10, dtype='float32')
-    >>> cspace = complex_space(rspace)
-    >>> cspace
-    uniform_discr(0.0, 1.0, 10, dtype='complex64')
-    >>> rspace.one().norm() == cspace.one().norm()
-    True
-    >>> complex_space(cspace) == cspace
-    True
-    """
-    from odl.discr.lp_discr import DiscreteLp
-    from odl.set.sets import ComplexNumbers
-    from odl.space.base_ntuples import FnBase, _TYPE_MAP_R2C
-    from odl.space.fspace import FunctionSpace
-
-    if isinstance(space, FunctionSpace):
-        return FunctionSpace(space.domain, field=ComplexNumbers())
-    else:
-        if is_real_floating_dtype(space.dtype):
-            complex_dtype = _TYPE_MAP_R2C[space.dtype]
-        elif is_complex_floating_dtype(space.dtype):
-            complex_dtype = space.dtype
-        else:
-            raise ValueError('data type {} is not a floating point type.'
-                             ''.format(space.dtype))
-
-    # DiscreteLp needs to come first since it inherits from FnBase
-    if isinstance(space, DiscreteLp):
-        return type(space)(complex_space(space.uspace), space.partition,
-                           dspace=complex_space(space.dspace),
-                           exponent=space.exponent, interp=space.interp,
-                           order=space.order)
-    elif isinstance(space, FnBase):
-        return type(space)(space.size, dtype=complex_dtype,
-                           weight=space._space_funcs)
-    else:
-        raise TypeError('space type {} not supported.'.format(type(space)))
-
-
-def real_space(space):
-    """Return the real counterpart of a given space.
-
-    Parameters
-    ----------
-    space : `LinearSpace`
-        Template space to be converted into its real counterpart.
-        If ``space`` is already a real space, a copy of it is returned.
-        Supported types of spaces: `FunctionSpace`, `FnBase`,
-        `DiscreteLp`.
-
-    Returns
-    -------
-    rspace : `LinearSpace`
-        Space of the same type and all parameters equal except field
-        and data type (if applicable), which are mapped to their real
-        counterparts.
-
-    Examples
-    --------
-    >>> from odl import uniform_discr
-    >>> cspace = uniform_discr(0, 1, 10, dtype='complex64')
-    >>> rspace = real_space(cspace)
-    >>> rspace
-    uniform_discr(0.0, 1.0, 10, dtype='float32')
-    >>> cspace.one().norm() == rspace.one().norm()
-    True
-    >>> real_space(rspace) == rspace
-    True
-    """
-    from odl.discr.lp_discr import DiscreteLp
-    from odl.set.sets import RealNumbers
-    from odl.space.base_ntuples import FnBase, _TYPE_MAP_C2R
-    from odl.space.fspace import FunctionSpace
-
-    if isinstance(space, FunctionSpace):
-        return FunctionSpace(space.domain, field=RealNumbers())
-    else:
-        if is_real_floating_dtype(space.dtype):
-            real_dtype = space.dtype
-        elif is_complex_floating_dtype(space.dtype):
-            real_dtype = _TYPE_MAP_C2R[space.dtype]
-        else:
-            raise ValueError('data type {} is not a floating point type.'
-                             ''.format(space.dtype))
-
-    # DiscreteLp needs to come first since it inherits from FnBase
-    if isinstance(space, DiscreteLp):
-        return type(space)(real_space(space.uspace), space.partition,
-                           dspace=real_space(space.dspace),
-                           exponent=space.exponent, interp=space.interp,
-                           order=space.order)
-    elif isinstance(space, FnBase):
-        return type(space)(space.size, dtype=real_dtype,
-                           weight=space._space_funcs)
-    else:
-        raise TypeError('space type {} not supported.'.format(type(space)))
-
-
 def preload_call_with(instance, mode):
     """Decorator to preload the first argument of a call method.
 
@@ -501,111 +362,6 @@ def preload_default_oop_call_with(vector):
         return oop_wrapper
 
     return decorator
-
-
-def fast_1d_tensor_mult(ndarr, onedim_arrs, axes=None):
-    """Fast multiplication of an n-dim array with an outer product.
-
-    This method implements the multiplication of an n-dimensional array
-    with an outer product of one-dimensional arrays, e.g.::
-
-        a = np.ones((10, 10, 10))
-        x = np.random.rand(10)
-        a *= x[:, None, None] * x[None, :, None] * x[None, None, :]
-
-    Basically, there are two ways to do such an operation:
-
-    1. First calculate the factor on the right-hand side and do one
-       "big" multiplication; or
-    2. Multiply by one factor at a time.
-
-    The procedure of building up the large factor in the first method
-    is relatively cheap if the number of 1d arrays is smaller than the
-    number of dimensions. For exactly n vectors, the second method is
-    faster, although it loops of the array ``a`` n times.
-
-    This implementation combines the two ideas into a hybrid scheme:
-
-    - If there are less 1d arrays than dimensions, choose 1.
-    - Otherwise, calculate the factor array for n-1 arrays
-      and multiply it to the large array. Finally, multiply with the
-      last 1d array.
-
-    The advantage of this approach is that it is memory-friendly and
-    loops over the big array only twice.
-
-    Parameters
-    ----------
-    ndarr : `numpy.ndarray`
-        Array to be multiplied with. Manipulations are done in-place.
-    onedim_arrs : sequence of array-like
-        One-dimensional arrays to be multiplied with ``ndarr``. The
-        sequence may not be longer than ``ndarr.ndim``.
-    axes : sequence of `int`, optional
-        Take the 1d transform along these axes. `None` corresponds to
-        the last ``len(onedim_arrs)`` axes, in ascending order.
-    """
-    if not isinstance(ndarr, np.ndarray):
-        raise TypeError('Expected a numpy.ndarray, got {!r}.'.format(ndarr))
-
-    if not onedim_arrs:
-        raise ValueError('No 1d arrays given.')
-
-    if axes is None:
-        axes = list(range(ndarr.ndim - len(onedim_arrs), ndarr.ndim))
-    elif len(axes) != len(onedim_arrs):
-        raise ValueError('There are {} 1d arrays, but {} axes entries.'
-                         ''.format(len(onedim_arrs), len(axes)))
-    else:
-        # Make axes positive
-        axes_ = np.array(axes, dtype=int)
-        axes_[axes_ < 0] += ndarr.ndim
-        axes = list(axes_)
-
-    if np.any(np.array(axes) >= ndarr.ndim) or np.any(np.array(axes) < 0):
-        raise ValueError('axes sequence contains out-of-bounds indices.')
-
-    # Make scalars 1d arrays and squeezable arrays 1d
-    alist = [np.atleast_1d(np.asarray(a).squeeze()) for a in onedim_arrs]
-    if any(a.ndim != 1 for a in alist):
-        raise ValueError('Only 1d arrays allowed.')
-
-    if len(axes) < ndarr.ndim:
-        # Make big factor array (start with 0d)
-        factor = np.array(1.0)
-        for ax, arr in zip(axes, alist):
-            # Meshgrid-style slice
-            slc = [None] * ndarr.ndim
-            slc[ax] = slice(None)
-            factor = factor * arr[slc]
-
-        ndarr *= factor
-
-    else:
-        # Hybrid approach
-
-        # Get the axis to spare for the final multiplication, the one
-        # with the largest stride.
-        axis_order = np.argsort(ndarr.strides)
-        last_ax = axis_order[-1]
-        last_arr = alist[axes.index(last_ax)]
-
-        # Build the semi-big array and multiply
-        factor = np.array(1.0)
-        for ax, arr in zip(axes, alist):
-            if ax == last_ax:
-                continue
-
-            slc = [None] * ndarr.ndim
-            slc[ax] = np.s_[:]
-            factor = factor * arr[slc]
-
-        ndarr *= factor
-
-        # Finally multiply by the remaining 1d array
-        slc = [None] * ndarr.ndim
-        slc[last_ax] = np.s_[:]
-        ndarr *= last_arr[slc]
 
 
 if __name__ == '__main__':
