+++ test/space/cu_ntuples_test.py
@@ -38,9 +38,6 @@ from odl.util.testutils import all_equal, all_almost_equal, almost_equal
 
 pytestmark = pytest.mark.skipif("not odl.CUDA_AVAILABLE")
 
-# TODO:
-# * custom dist/norm/inner
-
 
 # Helpers to generate data
